Time: 1 ms (95.3%), Space: 40.2 MB (18.34%) - LeetHub

+++ 3056-determine-if-a-cell-is-reachable-at-a-given-time/3056-determine-if-a-cell-is-reachable-at-a-given-time.java
@@ -1,15 +1,15 @@
 class Solution {
     public boolean isReachableAtTime(int sx, int sy, int fx, int fy, int t) {
-        final long startY = sy;
-        final long startX = sx;
-        final long endY = fy;
-        final long endX = fx;
-        final long dist = t;
-        final long distX = Math.abs(startX - endX);
-        final long distY = Math.abs(startY - endY);
-        final long distMax = Math.max(distX, distY);
+        final int startY = sy;
+        final int startX = sx;
+        final int endY = fy;
+        final int endX = fx;
+        final int dist = t;
+        final int distX = Math.abs(startX - endX);
+        final int distY = Math.abs(startY - endY);
+        final int distMax = Math.max(distX, distY);
         if (distMax == 0) {
-             return t !=1;
+            return t !=1;
         }
         return distMax <= t;
     }
